Fix unsafe logic in the heap
Constructors for heap objects (and also set_* operations) should not
take in a value without a lifetime - since we cannot assert that the
value is still valid.

So, I've renamed ReachableValue to Value, and it is the new user-facing
default. Our public APIs should now take in this new Value type.