Improve error reporting in the compiler.
Compiler errors are a major part of the user interface for a language.
Since I want to make it fun to program in Mold, I want to focus a LOT on
reporting errors that help a programmer narrow down the issue quickly. I
even want to provide "help" text whenever the compiler can identify
potential solutions.