Start working on the VM for Mold bytecode
This one was messy and involved changes in a lot of places. For
instance, I had to fix a bug with wrapping RootMarker in a RefCell for
the heap - RefCell doesn't Box, I didn't know that.

The VM can now push primitive values on the stack, and get and set
global variables.

I also added printing of runtime errors with a stack trace.